Prepared build infrastructure, shipping es, cjs and umd formats
