+++ .eslintrc
@@ -1,4 +1,6 @@
 ---
+  parserOptions:
+    sourceType: 'module'
   env:
     node: true
   rules:
