+++ .gitignore
@@ -1,4 +1,5 @@
 /artifacts/
 /build/
+dist
 node_modules
 npm-debug.log
