--- .npmignore
@@ -1,6 +1,0 @@
-.idea
-.coveralls.yml
-.eslintrc
-.travis.yml
-/artifacts/
-/tests/
