+++ package.json
@@ -2,16 +2,25 @@
   "name": "hoist-non-react-statics",
   "version": "2.4.0",
   "description": "Copies non-react specific statics from a child component to a parent component",
-  "main": "index.js",
+  "main": "dist/hoist-non-react-statics.cjs.js",
+  "module": "src/index.js",
   "types": "index.d.ts",
   "repository": {
     "type": "git",
     "url": "git://github.com/mridgway/hoist-non-react-statics.git"
   },
+  "files": [
+    "src",
+    "dist"
+  ],
   "scripts": {
     "cover": "node node_modules/istanbul/lib/cli.js cover --dir artifacts -- ./node_modules/mocha/bin/_mocha tests/unit/ --recursive --compilers js:babel/register --reporter spec",
-    "lint": "eslint ./index.js",
-    "test": "mocha tests/unit/ --recursive --compilers js:babel-register --reporter spec"
+    "lint": "eslint src",
+    "prebuild": "rimraf dist",
+    "build": "rollup -c",
+    "pretest": "npm run build",
+    "test": "mocha tests/unit/ --recursive --compilers js:babel-register --reporter spec",
+    "prepublish": "npm test"
   },
   "author": "Michael Ridgway <mcridgway@gmail.com>",
   "license": "BSD-3-Clause",
@@ -20,18 +29,20 @@
     "babel-cli": "^6.24.1",
     "babel-plugin-transform-class-properties": "^6.24.1",
     "babel-plugin-transform-react-jsx-source": "^6.22.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-es2016": "^6.24.1",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.24.1",
     "babel-register": "^6.24.1",
     "chai": "^4.0.1",
     "coveralls": "^2.11.1",
     "create-react-class": "^15.5.3",
-    "eslint": "^3.8.0",
+    "eslint": "^4.13.1",
     "istanbul": "^0.4.5",
     "mocha": "^3.4.2",
     "pre-commit": "^1.0.7",
-    "react": "^15.0.0"
+    "react": "^15.0.0",
+    "rimraf": "^2.6.2",
+    "rollup": "^0.52.3",
+    "rollup-plugin-uglify": "^2.0.1"
   },
   "keywords": [
     "react"
